fix: reduce main bundle size (exclude locale data)
By not loading locale data into the main process we can considerably
reduce it's size. We the only thing we need access to in the main
process is the translation sets (used for the languages menu).

refactor the i18n code so that the locale data can only be loaded when
needed (in the renderer).
